finished?
took the navbar collapse function out and then wanted to put it back in
but couldnt figure out how without redoing the whole navbar…

didnt know if i was sposed to add borders to the bottom boxes but when
i tried i couldnt get them to conform to the grid the way i wanted? did
i use rows right?